requirements.txt
anytree==2.12.1
chromadb==0.6.3
deap==1.4.2
fuzzywuzzy==0.18.0
httpx==0.28.1
keras_preprocessing==1.1.2
natsort==8.4.0
nltk==3.9.1
numpy==2.2.3
openai==1.66.3
pandas==2.2.3
psycopg2-binary
python_Levenshtein==0.27.1
recommonmark==0.7.1
regex==2024.11.6
Requests==2.32.3
scikit_learn==1.6.1
scipy==1.15.2
sentence_transformers==3.4.1
setuptools==76.0.0
sshtunnel==0.4.0
tenacity==9.0.0
textdistance==4.5.0
tiktoken==0.9.0
together==1.4.4
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
wordfreq==3.1.1
openpyxl
matplotlib
seaborn